+++ crates/control/src/button_filter.rs
@@ -1,8 +1,8 @@
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::Duration;
 
 use color_eyre::Result;
 use context_attribute::context;
-use filtering::tap_detector::TapDetector;
+use filtering::{debounce_button::DebounceButton, tap_detector::TapDetector};
 use framework::MainOutput;
 use serde::{Deserialize, Serialize};
 use types::{buttons::Buttons, cycle_time::CycleTime, sensor_data::SensorData};
@@ -81,48 +81,12 @@ impl ButtonFilter {
 
         Ok(MainOutputs {
             buttons: Buttons {
-                is_chest_button_pressed_once: self.chest_button_tap_detector.is_single_tapped,
+                is_chest_button_pressed_once: self.chest_button_tap_detector.is_single_tapped(),
                 head_buttons_touched: debounced_head_buttons_touched,
                 calibration_buttons_touched: debounced_calibration_buttons_touched,
                 animation_buttons_touched: debounced_animation_buttons_touched,
             }
             .into(),
         })
-    }
-}
-
-#[derive(Deserialize, Serialize)]
-struct DebounceButton {
-    last_button_touched: bool,
-    button_touched_time: SystemTime,
-}
-
-impl Default for DebounceButton {
-    fn default() -> Self {
-        Self {
-            last_button_touched: Default::default(),
-            button_touched_time: UNIX_EPOCH,
-        }
-    }
-}
-
-impl DebounceButton {
-    pub fn debounce_button(
-        &mut self,
-        button_touched: bool,
-        current_time: SystemTime,
-        timeout: Duration,
-    ) -> bool {
-        let button_touched_initially = button_touched && !self.last_button_touched;
-        if button_touched_initially {
-            self.button_touched_time = current_time;
-        }
-        self.last_button_touched = button_touched;
-
-        button_touched
-            && current_time
-                .duration_since(self.button_touched_time)
-                .unwrap()
-                >= timeout
     }
 }
