+++ crates/types/src/motion_selection.rs
@@ -51,7 +51,7 @@ impl Default for MotionType {
 #[derive(Clone, Debug, Deserialize, Serialize, PathSerialize, PathDeserialize, PathIntrospect)]
 pub struct MotionSafeExits {
     animation: bool,
-    animationstiff: bool,
+    animation_stiff: bool,
     arms_up_squat: bool,
     dispatching: bool,
     fall_protection: bool,
@@ -72,7 +72,7 @@ impl MotionSafeExits {
     pub fn fill(value: bool) -> Self {
         Self {
             animation: value,
-            animationstiff: value,
+            animation_stiff: value,
             arms_up_squat: value,
             dispatching: value,
             fall_protection: value,
@@ -95,7 +95,7 @@ impl Default for MotionSafeExits {
     fn default() -> Self {
         Self {
             animation: true,
-            animationstiff: true,
+            animation_stiff: true,
             arms_up_squat: true,
             dispatching: false,
             fall_protection: true,
@@ -120,7 +120,7 @@ impl Index<MotionType> for MotionSafeExits {
     fn index(&self, motion_type: MotionType) -> &Self::Output {
         match motion_type {
             MotionType::Animation => &self.animation,
-            MotionType::AnimationStiff => &self.animationstiff,
+            MotionType::AnimationStiff => &self.animation_stiff,
             MotionType::ArmsUpSquat => &self.arms_up_squat,
             MotionType::Dispatching => &self.dispatching,
             MotionType::Initial => &self.initial,
@@ -143,7 +143,7 @@ impl IndexMut<MotionType> for MotionSafeExits {
     fn index_mut(&mut self, motion_type: MotionType) -> &mut Self::Output {
         match motion_type {
             MotionType::Animation => &mut self.animation,
-            MotionType::AnimationStiff => &mut self.animationstiff,
+            MotionType::AnimationStiff => &mut self.animation_stiff,
             MotionType::ArmsUpSquat => &mut self.arms_up_squat,
             MotionType::Dispatching => &mut self.dispatching,
             MotionType::Initial => &mut self.initial,
